feat: preload dex cache server-side
