+++ apps/main/src/dex/components/PagePoolList/index.tsx
@@ -68,7 +68,7 @@ const PoolList = ({
   const isReady = useMemo(() => {
     // volume
     const haveVolumeMapper = typeof volumeMapper !== 'undefined' && Object.keys(volumeMapper).length >= 0
-    const volumeCacheOrApi = volumeMapper || volumeMapper || {}
+    const volumeCacheOrApi = volumeMapper || volumeMapperCached || {}
     const haveVolume = haveVolumeMapper || Object.keys(volumeCacheOrApi).length > 0
 
     // tvl
@@ -77,7 +77,7 @@ const PoolList = ({
     const haveTvl = haveTvlMapper || Object.keys(tvlCacheOrApi).length > 0
 
     return isLite ? haveTvl : haveVolume && haveTvl
-  }, [isLite, tvlMapper, tvlMapperCached, volumeMapper])
+  }, [isLite, tvlMapper, tvlMapperCached, volumeMapper, volumeMapperCached])
 
   const isReadyWithApiData = useMemo(() => {
     const haveVolume = typeof volumeMapper !== 'undefined' && Object.keys(volumeMapper).length >= 0
@@ -99,6 +99,12 @@ const PoolList = ({
     return keys.concat([COLUMN_KEYS.volume, COLUMN_KEYS.tvl])
   }, [isLite, isMdUp, showInPoolColumn])
 
+  console.log({
+    isReadyWithApiData,
+    noResult: formStatus.noResult,
+    isReady,
+    result: result?.length,
+  })
   const updateFormValues = useCallback(
     (searchParams: SearchParams) => {
       setFormValues(
